Now able to use socket: send txt file to server, and then server runs merlin
Addeed internet permission to use the socket

The socket is given as argument th IP address of my computer + a port: this port willbe the same on the JAVA SE program run on the PC

The SendToServer class takes a string (the one entered by user) as input.
Send it to the server. The part of the server code is alos coded in jave, but not android studio (using Intellij). I will try to publish it asap (I am able to run terminal commands, so it executes merlin :) )